evaluation/requirements.txt
# CLI Tool Evaluation Framework Dependencies
# Standalone requirements for the Docker-isolated CLI evaluation system

# Core dependencies
python-dotenv>=1.0.0    # .env file support for OAuth tokens
PyYAML>=6.0             # YAML parsing for scenario files

# Docker SDK (optional - for advanced Docker operations)
# docker>=6.0.0         # Uncomment if using Docker SDK instead of subprocess

# Development and testing (optional)
# pytest>=7.0.0         # For testing the evaluation framework
# pytest-cov>=4.0.0     # Test coverage reporting
